% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{AddNewClustertoDB}
\alias{AddNewClustertoDB}
\title{StoreProcessData}
\usage{
AddNewClustertoDB(
  cl = NULL,
  clusterName = NA,
  DBdir = .Options$RClusterWatcher.DBdir,
  verbose = FALSE
)
}
\arguments{
\item{cl}{parallel cluster object from parallel::makePSOCKcluster(names = 2)}

\item{clusterName}{character: optional name for cluster}

\item{DBdir}{database location defaults to .Options$RClusterWatcher.DBdir}

\item{verbose}{boolean, defaults to FALSE}
}
\value{
md5 heash location in database
}
\description{
StoreProcessData
}
\examples{
cl <- parallel::makePSOCKcluster(names = 2)
doParallel::registerDoParallel(cl)
ConnectClusterDB()
AddNewClustertoDB(cl)

#necessary to satisfy cran check
parallel::stopCluster(cl)
}
