% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedClusterFunctions.r
\name{AssesSlaveCluster}
\alias{AssesSlaveCluster}
\title{Asses state of a slave cluster}
\usage{
AssesSlaveCluster(RetrievedClusterList, verbose = FALSE)
}
\arguments{
\item{RetrievedClusterList}{output of Verify Cluster function @seealso StopSlaveClusters}

\item{verbose}{boolean be verbose or not}
}
\value{
an AssessedCluster Object
}
\description{
Asses state of a slave cluster
}
\examples{
\dontrun{
#internal function @seealso StopSlaveClusters
}
}
\keyword{internal}
