% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{CheckCluster}
\alias{CheckCluster}
\title{Check which parallel clusters are registerd in the database}
\usage{
CheckCluster(DBdir = .Options$RClusterWatcher.DBdir)
}
\arguments{
\item{DBdir}{database location defaults to .Options$RClusterWatcher.DBdir}
}
\value{
data.table containing the registered clusters
}
\description{
Check which parallel clusters are registerd in the database
}
\examples{
ConnectClusterDB()
CheckCluster()
}
