% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedClusterFunctions.r
\name{CheckforSlaveClusters}
\alias{CheckforSlaveClusters}
\title{Function to check if a Cluster also has attached slave clusters}
\usage{
CheckforSlaveClusters(
  MasterclusterHash,
  DBdir = .Options$RClusterWatcher.DBdir
)
}
\arguments{
\item{MasterclusterHash}{character clusterhash of a db}

\item{DBdir}{defaults to .Options$RClusterWatcher.DBdir}
}
\value{
the md5hash codes of these clusters in the database
}
\description{
Function to check if a Cluster also has attached slave clusters
}
\examples{
\dontrun{
 # internal function
 }
}
\keyword{internal}
