% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{ConnectClusterDB}
\alias{ConnectClusterDB}
\title{Create an Archivist DataBase to}
\usage{
ConnectClusterDB(DBdir = "C:", ResetDB = F)
}
\arguments{
\item{DBdir}{defaults to "C:"}

\item{ResetDB}{boolean to reset DataBase, Use True to reset DB, defaults to False}
}
\value{
invisible
}
\description{
Create an Archivist DataBase to
}
\examples{
# Make sure not to use a tempfile as in this example below
ConnectClusterDB(DBdir = dirname(tempfile()))

but rather:
\dontrun{
ConnectClusterDB(DBdir = "C:")
}

}
