% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Powershellfunctions.r
\name{GetProcessData}
\alias{GetProcessData}
\title{Function that Returns PID of current R session}
\usage{
GetProcessData(PID = Sys.getpid())
}
\arguments{
\item{PID}{process identification number, defaults to Sys.getpid()}
}
\value{
data.table with current PID, name and startup time
}
\description{
Currently only works on Windows
}
\examples{
GetProcessData()
}
