% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Powershellfunctions.r
\name{GetStartedClusterPIDData}
\alias{GetStartedClusterPIDData}
\title{Obtain ALL PID and startup information for a cluster}
\usage{
GetStartedClusterPIDData(cluster)
}
\arguments{
\item{cluster}{a cluster made by e.g. parallell:makecluster}
}
\value{
data.table with information about role and status of PID in a parallel cluster
}
\description{
Obtain ALL PID and startup information for a cluster
}
\examples{
\dontrun{
cl <- parallel::makePSOCKcluster(names = 2)
doParallel::registerDoParallel(cl)
GetStartedClusterPIDData(cl)
}
}
