% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Powershellfunctions.r
\name{KillPID}
\alias{KillPID}
\title{Kill a list of PIDs in Windows}
\usage{
KillPID(PIDlist)
}
\arguments{
\item{PIDlist}{array of PIDs that needs to be killed}
}
\value{
data.table with killed PID and there status
}
\description{
Kill a list of PIDs in Windows
}
\examples{
\dontrun{
cl <- parallel::makePSOCKcluster(names = 2)
doParallel::registerDoParallel(cl)
ClusterData <- GetStartedClusterPIDData(cl)


KillPID(ClusterData[Role == "worker",]$PID)
try(parallel::stopCluster(cl))}
}
