% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedClusterFunctions.r
\name{StopSlaveClusters}
\alias{StopSlaveClusters}
\title{Stop slave clusters of internal Clusters.}
\usage{
StopSlaveClusters(
  MasterclusterHash,
  DBdir = .Options$RClusterWatcher.DBdir,
  verbose = F
)
}
\arguments{
\item{MasterclusterHash}{ms5h hash of master cluster}

\item{DBdir}{defaults to .Options$RClusterWatcher.DBdir}

\item{verbose}{boolean defaults to FALSE}
}
\value{
nothing
}
\description{
Stop slave clusters of internal Clusters.
}
\examples{
\dontrun{
 #internal function
}
}
\keyword{internal}
