% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{TakeActionOnCluster}
\alias{TakeActionOnCluster}
\title{Once Verified and Assessed take Action on Cluster}
\usage{
TakeActionOnCluster(
  AssessedClusterObject,
  DBdir = .Options$RClusterWatcher.DBdir
)
}
\arguments{
\item{AssessedClusterObject}{output from AssesCluster function}

\item{DBdir}{database location defaults to .Options$RClusterWatcher.DBdir}
}
\value{
a list containing the KilledPids and the RemovedHash.
}
\description{
Once Verified and Assessed take Action on Cluster
}
\examples{
\dontrun{"internal function, dont use directly"}

}
\seealso{
VerifyCluster

AssesCluster
}
\keyword{Internal}
