% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{VerifyCluster}
\alias{VerifyCluster}
\title{Verify if a registered Cluster is also an active Cluster}
\usage{
VerifyCluster(ClusterHash = NULL, DBdir = .Options$RClusterWatcher.DBdir)
}
\arguments{
\item{ClusterHash}{md5 has of Database record}

\item{DBdir}{database location defaults to .Options$RClusterWatcher.DBdir}
}
\value{
verfiedClusterObject: list Containing, RegisterdCluster, ActiveCluster and md5 hash db location
}
\description{
Verify if a registered Cluster is also an active Cluster
}
\examples{
\dontrun{
# Make sure not to use a tempfile as in this example below
# ConnectClusterDB(DBdir = dirname(tempfile()))

# but rather e.g.:
\dontrun{
ConnectClusterDB(DBdir = "C:")
}
ConnectClusterDB(ResetDB = T, dirname(tempfile()))
cl <- parallel::makePSOCKcluster(names = 2)
doParallel::registerDoParallel(cl)
hash <- AddNewClustertoDB(cl)
test <- VerifyCluster(ClusterHash = hash)
}
}
\keyword{Internal}
