% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterfunctions.r
\name{AddNewClustertoDB}
\alias{AddNewClustertoDB}
\title{StoreProcessData}
\usage{
AddNewClustertoDB(
  cl = NULL,
  clusterName = "super cluster",
  MasterCluster = NA,
  DBdir = .Options$RClusterWatcher.DBdir,
  verbose = FALSE
)
}
\arguments{
\item{cl}{parallel cluster object from parallel::makePSOCKcluster(names = 2)}

\item{clusterName}{character: optional name for cluster}

\item{MasterCluster}{put in md5h hash of masterclusters, defaults to NA}

\item{DBdir}{database location defaults to .Options$RClusterWatcher.DBdir}

\item{verbose}{boolean, defaults to FALSE}
}
\value{
md5 heash location in database
}
\description{
StoreProcessData
}
\examples{
#' # Make sure not to use a tempfile as in this example below
# ConnectClusterDB(DBdir = dirname(tempfile()))

# but rather e.g.:
\dontrun{
ConnectClusterDB(DBdir = "C:")
}

cl <- parallel::makePSOCKcluster(names = 2)
doParallel::registerDoParallel(cl)
ConnectClusterDB(DBdir = dirname(tempfile()))
AddNewClustertoDB(cl)

#necessary to satisfy cran check
parallel::stopCluster(cl)
}
