% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{AssesCluster}
\alias{AssesCluster}
\title{Once verified,  Assess the health of cluster and return status and PIDs to be terminated}
\usage{
AssesCluster(RetrievedClusterList, verbose = F)
}
\arguments{
\item{RetrievedClusterList}{Output from VerifyCluster}

\item{verbose}{print output, defaults to FALSE}
}
\value{
a AssessedClusterObject with additional Assesment and PIDStoBeTerminated field
}
\description{
Once verified,  Assess the health of cluster and return status and PIDs to be terminated
}
\examples{
\dontrun{
ConnectClusterDB(ResetDB = T)
cl <- parallel::makePSOCKcluster(names = 2)
doParallel::registerDoParallel(cl)
hash <- AddNewClustertoDB(cl)
ClusterData <- VerifyCluster(ClusterHash = hash)
AssesCluster(ClusterData)
}
}
\seealso{
VerifyCluster
}
\keyword{Internal}
