% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterfunctions.r
\name{CheckCluster}
\alias{CheckCluster}
\title{Check which parallel clusters are registerd in the database}
\usage{
CheckCluster(DBdir = .Options$RClusterWatcher.DBdir, StopSlaveCls = T)
}
\arguments{
\item{DBdir}{database location defaults to .Options$RClusterWatcher.DBdir}

\item{StopSlaveCls}{Boolean, also try and stop slave clusters that are connected to a main slave cluster}
}
\value{
data.table containing the registered clusters
}
\description{
Check which parallel clusters are registerd in the database
}
\examples{
# Make sure not to use a tempfile as in this example below
# ConnectClusterDB(DBdir = dirname(tempfile()))

# but rather e.g.:
\dontrun{
ConnectClusterDB(DBdir = "C:")
}
ConnectClusterDB(dirname(tempfile()))
CheckCluster()
}
