% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist.r
\name{ConnectClusterDB}
\alias{ConnectClusterDB}
\title{Create an Archivist DataBase to}
\usage{
ConnectClusterDB(DBdir = "C://Database//dir", ResetDB = F)
}
\arguments{
\item{DBdir}{defaults to "C://Database//dir"}

\item{ResetDB}{boolean to reset DataBase, Use True to reset DB, defaults to False}
}
\value{
invisible
}
\description{
Create an Archivist DataBase to
}
\examples{
ConnectClusterDB()
}
